Save profile filename and Add and Delete groups
Each time a profile is saved or loaded, the filename will be saved and shown in the title of the window.

Added a new tab in the menu bar for "Groups". At the moment, empty groups can be added by going to "Group -> New -> Empty group".

Added a button to process groups to delete them.