Shortcuts for 'launch all' and 'stop all'
Added shortcuts for the main buttons of the process groups. To launch all, the shortcut will be `Alt+<group_num>` and `Alt+Shift+<group_num>` to stop all, where <group_num> is the position from left to right (counting from `1`).

Also, created a separate module for the class `AppWidget` that was in the `app_window.py` script before.