Fixed reference to the directory label
Fixed this bug: https://github.com/BorFour/process_launcher/issues/10
The bug was caused because the `Process` class had a direct reference to the `directory_widget`.  As this object changes in the `ProcessWidget` class, the reference was not updated.
Now, the `Process` class has a reference to its parent widget, so the reference is always updated.

Also, fixed a bug where the PID widget didn't update properly when change between modes