Fixed path issues for the images and installation of images
The path of the images of the icons is now relative to the python script that uses. 
The images have been moved inside of the `src/` folder, to they are inside the `process_launcher/` folder after the installation.
Modified the `setup.py` script to copy the images in the installation.